profiles_project/requirement.txt
Django==1.11
djangorestframework==3.6.2
pytz==2019.2
sqlparse==0.3.0
